+++ BeatBoxer-Engien/src/main/java/com/FerdiStro/drum/modes/pattern/objects/PatternIndex.java
@@ -1,0 +1,7 @@
+package com.FerdiStro.drum.modes.pattern.objects;
+
+import java.util.List;
+
+public record PatternIndex(
+        List<PatternMetaData> patterns
+) {}