+++ BeatBoxer-Engien/src/main/java/com/FerdiStro/drum/modes/pattern/objects/PatternMetaDataModes.java
@@ -1,0 +1,5 @@
+package com.FerdiStro.drum.modes.pattern.objects;
+
+public enum PatternMetaDataModes {
+    OFFLINE, CONNECT
+}