+++ BeatBoxer-Engien/src/main/java/com/FerdiStro/drum/modes/pattern/ques/SearchQue.java
@@ -1,0 +1,7 @@
+package com.FerdiStro.drum.modes.pattern.ques;
+
+public enum SearchQue {
+    ALL,
+    MODE,
+
+}